chore: rename analyzer utils for clarity
