+++ src/analyzer/typescript-utils/index.ts
@@ -1,2 +1,3 @@
 export * from './typescript-export';
 export * from './typescript-type';
+export * from './typescript-utils';
