+++ src/analyzer/typescript-utils/typescript-internals.d.ts
