+++ src/analyzer/typescript-utils/typescript-utils.ts
